Refactor WebhookVerifier interface w/ Danny
