+++ packages/api/tsconfig.json
@@ -9,7 +9,7 @@
       "src/*": ["./src/*"]
     }
   },
-  "include": ["src", "src/../package.json"],
+  "include": ["src/**/*", "src/../package.json"],
   "references": [
     { "path": "../internal" },
     { "path": "../auth" },
